refactor(store): split clickhouse.go into per-responsibility files
What:
- split internal/store/clickhouse.go (3429 lines) into 6 files within
  the same `store` package:
  - clickhouse.go        common: config, OpenConn, retryFlush, dynCfg
  - clickhouse_trace.go  ClickHouseTraceStore + QueryInfraCorrelation
  - clickhouse_metric.go ClickHouseMetricStore
  - clickhouse_log.go    ClickHouseLogStore
  - clickhouse_schema.go ensure*Table DDL
  - clickhouse_util.go   OTel attribute helpers
- pure move: no logic, signature, or API change

Why:
- 3429-line file hurt navigation/review; per-responsibility files
  make the trace/metric/log stores and schema easier to work on

Verify:
- go build ./... / go vet ./... / gofmt -l . (clean)
- go test -race -count=1 ./... (all pass)
- 68 top-level declarations identical to original (diff)
- 3136 non-blank body lines byte-identical to original (diff);
  only gofmt blank-line spacing differs
